+++ src/Connection.php
@@ -10,10 +10,24 @@ namespace atk4\dsql;
  */
 class Connection
 {
+    /** @var string Query classname */
     protected $query_class = 'atk4\dsql\Query';
+    
+    /** @var string Expression classname */
     protected $expression_class = 'atk4\dsql\Expression';
+    
+    /** @var Connection Connection object */
     protected $connection = null;
 
+    /**
+     * Connect database
+     *
+     * @param string $dsn
+     * @param string $user
+     * @param string $password
+     * @param array $args
+     * @return Connection
+     */
     public static function connect($dsn, $user = null, $password = null, $args = [])
     {
         if (strpos($dsn, ':') === false) {
@@ -24,28 +38,28 @@ class Connection
         switch (strtolower($driver)) {
             case 'mysql':
                 return new Connection(array_merge([
-                    'connection'=>new \PDO($dsn, $user, $password),
-                    'query_class'=>'atk4\dsql\Query_MySQL'
+                    'connection' => new \PDO($dsn, $user, $password),
+                    'query_class' => 'atk4\dsql\Query_MySQL'
                 ], $args));
             case 'sqlite':
                 return new Connection(array_merge([
-                    'connection'=>new \PDO($dsn, $user, $password),
-                    'query_class'=>'atk4\dsql\Query_SQLite'
+                    'connection' => new \PDO($dsn, $user, $password),
+                    'query_class' => 'atk4\dsql\Query_SQLite'
                 ], $args));
             case 'dumper':
                 return new Connection_Dumper(array_merge([
-                    'connection'=>Connection::connect($rest)
+                    'connection' => Connection::connect($rest)
                 ], $args));
 
             case 'counter':
                 return new Connection_Counter(array_merge([
-                    'connection'=>Connection::connect($rest)
+                    'connection' => Connection::connect($rest)
                 ], $args));
 
                 // let PDO handle the rest
             default:
                 return new Connection(array_merge([
-                    'connection'=>new \PDO($dsn, $user, $password)
+                    'connection' => new \PDO($dsn, $user, $password)
                 ], $args));
 
         }
@@ -55,11 +69,11 @@ class Connection
      * Specifying $attributes to constructors will override default
      * attribute values of this class.
      *
-     * @param array        $attributes
+     * @param array $attributes
      */
     public function __construct($attributes = null)
     {
-        if ($attributes) {
+        if ($attributes !== null) {
             if (!is_array($attributes)) {
                 throw new Exception('Invalid arguments for "new Connection()". Did you mean to call Connection::connect()?');
             }
@@ -70,6 +84,12 @@ class Connection
         }
     }
 
+    /**
+     * Returns new Query object with connection already set
+     *
+     * @param array $properties
+     * @return Query
+     */
     public function dsql($properties = [])
     {
         $c = $this->query_class;
@@ -79,6 +99,13 @@ class Connection
         return $q;
     }
 
+    /**
+     * Returns Expression object with connection already set
+     *
+     * @param array $properties
+     * @param array $arguments
+     * @return Expression
+     */
     public function expr($properties = [], $arguments = null)
     {
         $c = $this->expression_class;
@@ -88,11 +115,22 @@ class Connection
         return $e;
     }
 
+    /**
+     * Returns Connection object
+     *
+     * @return Connection
+     */
     public function connection()
     {
         return $this->connection;
     }
 
+    /**
+     * Execute Expression by using this connection
+     *
+     * @param Expression $expr
+     * @return PDOStatement
+     */
     public function execute(Expression $expr)
     {
         // If custom connection is set, execute again using that