+++ tests/ExpressionTest.php
@@ -4,6 +4,7 @@ namespace atk4\dsql\tests;
 
 use atk4\dsql\Expression;
 use atk4\dsql\Expressionable;
+use atk4\dsql\Query;
 
 /**
  * @coversDefaultClass \atk4\dsql\Expression
@@ -258,6 +259,18 @@ class ExpressionTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
+     * @covers ::__toString
+     * @expectedException \Exception
+     */
+    /*
+    public function testToStringException1()
+    {
+        $e = new MyBadExpression('Hello');
+        $s = (string)$e;
+    }
+    */
+
+    /**
      * expr() should return new Expression object and inherit connection from it.
      *
      * @covers ::expr
@@ -272,6 +285,7 @@ class ExpressionTest extends \PHPUnit_Framework_TestCase
      * Fully covers _escape method
      *
      * @covers ::_escape
+     * @covers ::_escapeSoft
      * @covers ::escape
      */
     public function testEscape()
@@ -354,7 +368,6 @@ class ExpressionTest extends \PHPUnit_Framework_TestCase
     public function testParam()
     {
         $e = new Expression('hello, [who]', ['who' => 'world']);
-
         $this->assertEquals(
             'hello, :a',
             $e->render()
@@ -364,14 +377,16 @@ class ExpressionTest extends \PHPUnit_Framework_TestCase
             $e->params
         );
 
-        $e = new Expression('hello, [who]', ['who' => 'world']);
-
+        // @todo Imants: allowing to pass value as array looks wrong.
+        //      See test case in testParam() method.
+        //      Maybe we should add implode(' ', array_map(...)) here ?
+        $e = new Expression('hello, [who]', ['who' => ['cruel', 'world']]);
         $this->assertEquals(
-            'hello, :a',
+            'hello, (:a,:b)',
             $e->render()
         );
         $this->assertEquals(
-            [':a'=>'world'],
+            [':a'=>'cruel', ':b'=>'world'],
             $e->params
         );
     }
@@ -393,6 +408,10 @@ class ExpressionTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(
             123,
             PHPUnitUtil::callProtectedMethod($this->e(), '_consume', [123, 'none'])
+        );
+        $this->assertEquals(
+            '(select *)',
+            PHPUnitUtil::callProtectedMethod($this->e(), '_consume', [new Query()])
         );
 
         $this->assertEquals(
@@ -500,6 +519,18 @@ class ExpressionTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
+     * Test var-dump code for codecoverage
+     *
+     * @covers ::__debugInfo
+     */
+    public function testVarDump()
+    {
+        $this->e('test')->__debugInfo();
+
+        $this->e(' [nosuchtag] ')->__debugInfo();
+    }
+
+    /**
      * Test reset()
      *
      * @covers ::reset
@@ -535,10 +566,21 @@ class JsonExpression extends Expression
         return json_encode($value);
     }
 }
-class MyField implements Expressionable {
-    function getDSQLExpression($e)
+class MyField implements Expressionable
+{
+    public function getDSQLExpression($e)
     {
         return $e->expr('`myfield`');
     }
 }
+/*
+class MyBadExpression extends Expression
+{
+    public function getOne()
+    {
+        // should return string, but for test case we return array to get \Exception
+        return array();
+    }
+}
+*/
 // @codingStandardsIgnoreEnd